Move User helper methods
